+++ tests/phpunit/integration/hypeJunction/Discovery/HooksTest.php
@@ -2,6 +2,7 @@
 
 namespace hypeJunction\Discovery;
 
+use Elgg\Event;
 use Elgg\IntegrationTestCase;
 
 /**
@@ -18,15 +19,25 @@ class HooksTest extends IntegrationTestCase {
 
 	public function down() {}
 
+	private function makeEvent($value, array $params = []): Event {
+		$event = $this->getMockBuilder(Event::class)->disableOriginalConstructor()->getMock();
+		$event->method('getValue')->willReturn($value);
+		$event->method('getParams')->willReturn($params);
+		$event->method('getParam')->willReturnCallback(function (string $key, $default = null) use ($params) {
+			return array_key_exists($key, $params) ? $params[$key] : $default;
+		});
+		return $event;
+	}
+
 	public function testPublicPagesHookAddsPermalinkAndShareAction(): void {
-		$return = Router::publicPages('public_pages', 'walled_garden', []);
+		$return = Router::publicPages($this->makeEvent([]));
 		$this->assertIsArray($return);
 		$this->assertContains('permalink/.*', $return);
 		$this->assertContains('action/discovery/share', $return);
 	}
 
 	public function testOpenGraphImageSizesHookReturnsSizes(): void {
-		$return = Icons::entityOpenGraphImageSizes('entity:open_graph_image:sizes', 'object', [], []);
+		$return = Icons::entityOpenGraphImageSizes($this->makeEvent([]));
 		$this->assertIsArray($return);
 		$this->assertArrayHasKey('large', $return);
 		$this->assertArrayHasKey('medium', $return);
@@ -38,10 +49,7 @@ class HooksTest extends IntegrationTestCase {
 	public function testGraphExportHookReturnsSiteTags(): void {
 		$site = elgg_get_site_entity();
 		$return = Discovery::graphExport(
-			'metatags',
-			'discovery',
-			[],
-			['entity' => $site, 'url' => $site->getURL()]
+			$this->makeEvent([], ['entity' => $site, 'url' => $site->getURL()])
 		);
 		$this->assertArrayHasKey('og:type', $return);
 		$this->assertEquals('website', $return['og:type']);
@@ -50,21 +58,19 @@ class HooksTest extends IntegrationTestCase {
 
 	public function testEntityMenuRegisterHookRuns(): void {
 		$user = $this->createUser();
-		elgg_get_session()->setLoggedInUser($user);
+		_elgg_services()->session_manager->setLoggedInUser($user);
 		try {
-			// Trigger should not fatal; menu items may or may not be added
-			$result = elgg_trigger_plugin_hook('register', 'menu:entity', [
-				'entity' => elgg_get_site_entity(),
-			], []);
+			$entity = elgg_get_site_entity();
+			$result = Menus::entityMenuSetup($this->makeEvent([], ['entity' => $entity]));
 			$this->assertIsArray($result);
 		} finally {
-			elgg_get_session()->removeLoggedInUser();
+			_elgg_services()->session_manager->removeLoggedInUser();
 		}
 	}
 
 	public function testRedirectErrorToPermalinkReturnsOriginalForUnknownUrl(): void {
 		$return = 'http://example.com/original';
-		$result = Router::redirectErrorToPermalink('forward', '404', $return, []);
+		$result = Router::redirectErrorToPermalink($this->makeEvent($return, []));
 		$this->assertIsString($result);
 	}
 }